Day 1
Notes:
Using unwrap() to deal with errors because it's the easiest to program and therefore the fastest for getting a solution to the puzzle.
For the same reason, code is read to string and kept in memory rather than using a lazy reader.
For part 2, the sum of the sweeping windows does not need to be calculated because of the overlap between the two windows.