day 4
Notes:
Complexity is O(draws^2), which could be O(draws) if we store a state with the boards by marking (or unmarking for part2) cells when iterating through the draws.
Dealing with the board_win would be easier with 2D arrays (or nested vectors), but the  types and parsing would be more annoying to deal with.