day 9
Notes:
Exactly as I had it at the time of the solve. get_adj took me longer to write than I'm happy with... Vec<_> should just have a standard library solution for indexing with a potentially negative number. Probably checking the size would be cleaner in the end. I'm also mixing iterators and vectors. Probably this would be better with just iterators.