day 10
Fairly simple problem, Rust pattern matching is amazing for this. For cleanup there was the decision of where to put the mappings from character to score. At first I had a separate function for the mapping, but this version is more concise and still easily readable. Also there is always a choice between using iterators or for-loops. Usually simple code is nicer with iterators (like in part1), but nested closures can become difficult to read (or debug).